requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
yfinance==0.2.58
pandas==2.2.3
numpy==2.2.5
sqlalchemy==2.0.41
PyJWT==2.10.1
